Script changes
Added the new version of the AI that the programmers made. Created a default interactions script for the doors, so we can use the same one over and over. tried to make level select work. created a levelspawn script that will take care of spawning all the right layouts on the right levels.