Searchable Object Prefab
Removed all item tags from objects.
Created new prefab that is the only one that can be searched. It works in Edward's scene, but is a little buggy since the objects are so close together. Still have an error in Alcimus. (Edward and Alcimus are the only ones with updated maps)